+++ test/unixtcpsrvr.lua
@@ -1,6 +1,6 @@
     socket = require"socket"
     socket.unix = require"socket.unix"
-    u = assert(socket.unix())
+    u = assert(socket.unix.tcp())
     assert(u:bind("/tmp/foo"))
     assert(u:listen())
     c = assert(u:accept())
